image-server/requirements.txt
torch
diffusers[torch]
transformers
accelerate
Pillow
fastapi
uvicorn
uvicorn[standard]
python-dotenv
torchvision
sse-starlette
watchgod